pyproject.toml
[project]
name = "book-service"
version = "0.1.0"
description = "book services dependencies"
authors = [
    {name = "olumide osho"}
]
package-mode = false
#readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pymongo (>=4.11.1,<5.0.0)",
    "motor (>=3.7.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "jwt (>=1.3.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "confluent-kafka (>=2.8.0,<3.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "boto3 (>=1.36.25,<2.0.0)",
    "django-storages (>=1.14.5,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-filter (>=25.1,<26.0)",
    "cuid2 (>=2.0.1,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "drf-standardized-errors (>=0.14.1,<0.15.0)",
    "pillow (>=11.1.0,<12.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
